# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "DDD-based file storage management project built with Gin.",
        "title": "Power File Storage API",
        "contact": {},
        "version": "1.0"
    },
    "paths": {
        "/api/fs": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "fs"
                ],
                "summary": "创建文件或目录",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/mixin.CreateCaseRes"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/woosh.ProblemDetails"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/woosh.ProblemDetails"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/woosh.ProblemDetails"
                        }
                    }
                }
            },
            "delete": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "fs"
                ],
                "summary": "移除文件或目录",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/mixin.DeleteCaseRes"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/woosh.ProblemDetails"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/woosh.ProblemDetails"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/woosh.ProblemDetails"
                        }
                    }
                }
            }
        },
        "/api/fs/move": {
            "put": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "fs"
                ],
                "summary": "移动文件或目录",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/mixin.MoveCaseRes"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/woosh.ProblemDetails"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/woosh.ProblemDetails"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/woosh.ProblemDetails"
                        }
                    }
                }
            }
        },
        "/api/fs/rename": {
            "put": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "fs"
                ],
                "summary": "重命名文件或目录",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/mixin.RenameCaseRes"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/woosh.ProblemDetails"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/woosh.ProblemDetails"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/woosh.ProblemDetails"
                        }
                    }
                }
            }
        },
        "/api/fs/{id}": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "fs"
                ],
                "summary": "获取文件或目录的详细信息",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "操作id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/mixin.GetInfoCaseRes"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/woosh.ProblemDetails"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/woosh.ProblemDetails"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/woosh.ProblemDetails"
                        }
                    }
                }
            }
        },
        "/api/fs/{id}/directories": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "fs"
                ],
                "summary": "获取指定目录下一层级的目录列表",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "操作id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "搜索",
                        "name": "search",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "页码",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "页的大小",
                        "name": "pageSize",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/mixin.GetDirectoriesCaseRes"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/woosh.ProblemDetails"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/woosh.ProblemDetails"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/woosh.ProblemDetails"
                        }
                    }
                }
            }
        },
        "/api/fs/{id}/files": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "fs"
                ],
                "summary": "获取指定目录下一层级的文件列表",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "操作id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "搜索",
                        "name": "search",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "页码",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "页的大小",
                        "name": "pageSize",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/mixin.GetDirectoriesCaseRes"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/woosh.ProblemDetails"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/woosh.ProblemDetails"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/woosh.ProblemDetails"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "mixin.CreateCaseRes": {
            "type": "object"
        },
        "mixin.DeleteCaseRes": {
            "type": "object"
        },
        "mixin.GetDirectoriesCaseRes": {
            "type": "object",
            "properties": {
                "hasNextPages": {
                    "type": "boolean",
                    "format": "bool",
                    "example": true
                },
                "hasPrevPages": {
                    "type": "boolean",
                    "format": "bool",
                    "example": true
                },
                "items": {
                    "type": "array",
                    "items": {
                        "format": "[]GetDirectoryCaseRes",
                        "$ref": "#/definitions/mixin.GetDirectoryCaseRes"
                    }
                },
                "pageIndex": {
                    "type": "integer",
                    "format": "int",
                    "example": 1
                },
                "pageSize": {
                    "type": "integer",
                    "format": "int",
                    "example": 10
                },
                "totalCount": {
                    "type": "integer",
                    "format": "int",
                    "example": 50
                },
                "totalPages": {
                    "type": "integer",
                    "format": "int",
                    "example": 5
                }
            }
        },
        "mixin.GetDirectoryCaseRes": {
            "type": "object"
        },
        "mixin.GetInfoCaseRes": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer",
                    "format": "uint64",
                    "example": 1
                },
                "isDirectory": {
                    "type": "boolean",
                    "format": "bool",
                    "example": true
                },
                "name": {
                    "type": "string",
                    "format": "string",
                    "example": "a new dir"
                }
            }
        },
        "mixin.MoveCaseRes": {
            "type": "object"
        },
        "mixin.RenameCaseRes": {
            "type": "object"
        },
        "woosh.ProblemDetails": {
            "type": "object",
            "properties": {
                "detail": {
                    "type": "string",
                    "format": "string",
                    "example": "a error detail"
                },
                "instance": {
                    "type": "string",
                    "format": "string",
                    "example": "/api/to/path"
                },
                "status": {
                    "type": "integer",
                    "format": "int",
                    "example": 400
                },
                "title": {
                    "type": "string",
                    "format": "string",
                    "example": "a error title"
                }
            }
        }
    }
}